+++ src/main/java/it/unipi/booknetapi/model/book/BookRecommendation.java
@@ -8,13 +8,13 @@ import lombok.*;
 @EqualsAndHashCode(callSuper = true)
 public class BookRecommendation extends BookEmbed {
 
-    private Long score;
+    private Double score;
 
     public BookRecommendation(Book book) {
         super(book);
     }
 
-    public BookRecommendation(Book book, Long score) {
+    public BookRecommendation(Book book, Double score) {
         super(book);
         this.score = score;
     }
